Convert enum to front/back, fix async bugs
